feat: 6. 그래프 관련 기능 통합
